Removed field "dtCtors :: [Ctor n t]" from type "DataType n t".
This change moves toward an acyclic term structure, where the shared
terms are dags instead of cyclic graphs. It should be possible to
traverse terms using primitive recursion, and to construct shared
terms in the IO monad without needing MonadFix or similar to create
cycles. (We will also need to remove field "defEqs" from type "Def e"
before this is completely true.)

The list of constructors for each datatype is now stored in the
Module structure, as part of the "moduleTypeMap" table. A list of
constructors is now included explicitly in the type synonym
"TypedDataType".