Add little-endian versions of split and join.
The arrays in our ECDSA code are little-endian (that is, the least significant part of a 384-bit word is stored at index 0 in an array). Previously, Cryptol's split and join were aligned with this interpretation. Cryptol v2 does split and join on bit vectors in the other order, however, so we need a special operation to deal with target programs that model large integers as arrays with the least significant element first.