Implement java_symexec
This is like java_extract, but takes a list of inputs and a list of outputs
instead of a JavaSetup block. Additionally, it takes an initial binding for
each input, which can be the result of `fresh_bitvector` or the like. The
resulting term will then contain external constants, which can be used directly
or abstracted using `abstract_symbolic`. Inputs and outputs can be scalars or
arrays of scalars. Full objects, and arrays of objects, will come soon.

This also enables the return value of `extract_java` to be an array, though
the return value is still the only output in models generated using that
primitive.