Reimplement '*PrintPrim' in terms of '*Prim'.
This adds a missing 'checkBooleanSchema' check to 'satPrintPrim', the
lack of which looked like a bug. However, the 'yices' and 'abc'
provers already embed a 'checkBooleanSchema' check, so the missing
check may not have mattered in practice.

I wanted to implement 'prove_print' and 'sat_print' in SAWScript using
bind and 'print'. But this is difficult, because 'prove_print' is
also, *mysteriously*, the only way to produce a 'Theorem'! On the
other hand, non-symmetrically, 'sat_print' really is just 'sat'
followed by 'print'. It doesn't make sense to simply change
'prove_print' to symmetrically be 'prove' followed by 'print', since
then the user would have a 'ProofResult' when they really wanted a
'Theorem'. The 'Theorem' type is used with 'addsimp' (only used in a
few examples). Assuming that it doesn't make sense to produce theorems
at will in the common case -- an axiom backdoor like 'assume' or
'axiom' for producing 'Theorem's at will makes sense for the uncommon
case -- we need a privileged operation like 'prove_print' that
produces verified theorems.

One option is to change the 'ProofResult' type to embed a
theorem. I.e. instead of

  data ProofResult
    = Valid
    | Invalid FiniteValue
    | InvalidMulti [(String, FiniteValue)]

to

  data ProofResult
    = Valid (Theorem SAWCtx)
    | Invalid FiniteValue
    | InvalidMulti [(String, FiniteValue)]

and then the magic 'Theorem' production in 'prove_print' could be
replaced by use of 'caseProofResult', and maybe

  proofResultToTheorem :: ProofResult -> Theorem SAWCtx
  proofResultToTheorem (Valid t) = t
  proofResultToTheorem r = fail (show r)

as a more direct replacement for the current "prove or die" behavior
of 'prove_print'. Another option would be a 'prove_or_die' primitive:

  proveOrDiePrim _sc script t =
    proofResultToTheorem =<< provePrim _sc script t

which is what 'prove_print' is currently doing, besides printing in
the success case.

One problem with changing the definition of 'ProofResult' is that
'SatResult' is currently symmetric to 'ProofResult',

  data SatResult
    = Unsat
    | Sat FiniteValue
    | SatMulti [(String, FiniteValue)]

and it doesn't make sense to change the 'Unsat' constructor to 'Unsat
(Theorem SAWCtx)' unless we have a way to negate SAW terms. Currently,
the negation appears to happen at the level of AIGs (using 'AIG.not').