Add dummy PIs to fib examples and add buggy fib example.
In the new, buggy fib example, 'fib3', the state machine returns the
wrong output when an added state counter has a magic value
(0xff00ff00ff00ff00, more than 2^63). After 3 or 4 hours of waiting,
ABC's sequential equivalence checker 'dsec' is unable to find the
bug. The trouble may be that ABC needs to check that the bad state is
reachable, and without an inductive argument showing that all counter
values are reachable, ABC will need more than 2^63 simulation steps to
reach the magic counter value.

The dummy PIs could be added programmatically, but I doubt this is a
common case. On the ABC command line they can be added like this:

  abc> read foo.aig
  abc> addpi
  abc> write foo.addpi.aig

The wrong way to add a dummy PI programmatically is to use the
'ABC.newInput' function from abc-bridge. The problem is that the
position of the latches is fixed as the last 'k' bits, for some user
specified 'k', and the 'ABC.newInput' command doesn't let you specify
*where* the new input should go (and it doesn't go at the beginning in
my experiments).  A simple way to add a dummy PI at the beginning, pre
bit blast, is to add an ignored input to the SAW term (like wrapping
in 'const'). For a SAW term 't', we transform

  t |-> (\(_:[1]) -> t)

I have not actually implemented this though, because I expect the case
of no inputs is rare (my Fibonacci examples are contrived).