WIP: Change bit-blasting interface to avoid bit-engine-reuse errors.
The LSS and JSS SAW backends were preserving a bit engine and reusing
it in each bit blast. This led to duplicated inputs in AIGs -- and
CNFs, but that did not affect satisfiability -- whenever more than one
was created by the LSS or JSS run. So, I changed the bit-blaster
interfaces to prevent this erroneous usage: they now create their own
bit-engine internally.