SAWScript interpreter: allow non-monadic values as toplevel statements
If the expression in a toplevel statement (or typed at the REPL) does
not have a TopLevel monadic type, then it behaves as if there were an
implicit "return" applied to it, instead of reporting an error message.
At the REPL, the effect is that the value gets printed out (as long as
it is not of unit type).