Add sizes to llvm_symexec inputs and outputs
Each input and output entry in the parameters to llvm_symexec now takes a
count of how many elements to read or write at the given location.

For inputs, this could be inferred from the passed value. For outputs, it
could be inferred from allocation sizes, but passing it in explicitly
provides more flexibility.