Add summary graphs of scaling experiments.
Not getting the results I expected here. Not included is iimc being unable
to prove equivalence of ECB when scaling by 85, but doing fine up to 80. Here
we have ABC 'dprove' on the ECB miter, and ABC 'cec' comparing the circuits.
The run time doesn't correlate much with the circuit size, but it's also not
blowing up as bad as I expected, given that the AES example takes more than
20 hours (I killed it) for the ECB miter, whereas it takes 4 to 15 minutes
for the corresponding 'cec' comparison.