Add scaling test results and change y-axis transform on some graphs.
Switched timing graph scales from log to linear.

Added timing results for ABC 'dsec' and '&cec'.

I'm not concluding much from these results. The AES results indicated
that verifying cipher modes this way was not going to work, since
nothing I tried would verify ECB AES in a reasonable amount of time,
relative to the time it took to compare the two AES implementations
directly. But for these artificial examples, comparing the scaling
implementations, the time to compare the ECB circuits is *faster* than
the time to compare the underlying block "ciphers" directly!

Also, 'cec', 'dsec', and 'dprove' result graphs have similar shapes,
but the slowest succeeding tests (65, 70, and 75) for '&cec' are some
of fastest for the other commands. This could be due to the 'cec',
'dsec', and 'dprove' implementations sharing a lot of code, whereas
'&cec' is a "new" command and hence possibly quite different.