Use new Cryptol expression defaulter.
The old defaulter, 'Cryptol.TypeCheck.Defaulting.defaultExpr', was
removed from Cryptol recently
(58bb7254da0db006bd91720a1c53bc3a2bc37d80). To use the new defaulter,
I just copied the changes that happened in Cryptol itself. The CVC4
'SolverConfig' from a test 'Cryptol.TypeCheck.Solve'; in a running
Cryptol program this config is read from the environment, but I wasn't
sure how to do that in saw-script, and doubt that it matters (we
assume CVC4 as a dependency anyway).