Convert SAWCore to use nested representation for tuples.
The core term representation now uses Unit and binary Pair
values/types to represent tuples of all sizes. Eventually
this will make it possible to reduce our reliance on custom
rewriter conversions and fancy polymorphic primitives for
dealing with tuples and records.

The new Pair value constructor is printed as infix "x # y",
and the Pair type constructor is printed as "a * b". Left and
right projections are printed as "x.L" and "x.R". Currently
this syntax is only used for output.

Tuple values like "(x, y, z)" are encoded as "x # (y # (z # ()))".
Tuple types like "#(a, b, c)" are encoded as "a * (b * (c * #()))".
Tuple selectors like "x.2" are encoded as "((x.R).R).L".

There are more steps yet to be done:
- Use nested representation also for records
- Fix pretty printer to recover standard tuple syntax
- Fix parser to allow writing low-level Pair values/types/selectors