Extend saw-core parser with syntax for nested-record constructors
The syntax "{ label = x | r }" denotes the record "r" extended with
a field named "label" with value "x". Similarly, "#{ label :: a | t }"
denotes the record type "t" extended with a field named "label" of type
"a".

A parenthesized expression of type String may be used in place of a
label, such as "{ (s) = x | r }". This makes it possible to dynamically
construct records with arbitrary field names. This syntax can also be
used with ordinary record expressions like "{ (s) = x, b = y }" or with
record selectors like "r.(s)".