Cryptol import now uses KType datatype and "ty" function
KType and ty were removed in the "native tuples/records" changeset
from Jan 27. This patch partially reverts that changeset.
However, due to the new nested implementation of tuples/records in
saw-core, the new implementation of "ty" is now able to use "native"
tuples and records.

As of this changeset we are still using the overloaded primitives
(unary, binary, eq, zero, etc.) but it should now be possible to
reintroduce pattern matching definitions to make those primitives
unnecessary.