Update scWhnf to deal appropriately with nested tuples/records
The underlying Pair and Field data constructors are now considered
to be tail-strict: scWhnf will force the evaluation of the entire
spine of any tuple or record. Tuple selectors are now handled at the
level of the underlying Pair selectors. For record selectors, we now
force the spine and labels of the record before attempting to look up
the selected label.