Change definition of "pmul" back to use little-endian operations
This reverts part of changeset f3a7ebd1. Function "pmul" actually
gives the same result with either bit order, but the little-endian
versions avoid gratuitous applications of "reverse", and make the
definition of "pmul" consistent with "pdiv" and "pmod".

The problem described in GaloisInc/saw-script#101 was due to a
buggy implementation of "blastLE", not the endianness of the definition
of "pmul".