Assume ptrs allocated in llvm_verify are non-null
Given that they're considered "valid", they should be assumed to be
non-null.

Similarly, we should add an assumptions that any allocated object has an
address less than the maximum address minus the size of the object, so
that none of the offset addresses inside the object can be zero due to
address arithmetic wrapping around. This commit does not make that
change, however.