All multiple overrides for one LLVM function
Now you can pass multiple method specs corresponding to a single LLVM
function into `llvm_verify`. If it is successful in running exactly one
of them, it will use that result. Otherwise it will fail. If
`llvm_sat_branches` is specified, then it will perform a SAT check on
any assertions in the process of determining which override to use. At
the moment, it tends to be rarely useful without `llvm_sat_branches`
enabled.