examples: add ChaCha20
This is the same code example I gave in my Compose Conference 2016
talk, along with a simple Makefile for verification. It's also been
slightly modified for rigid RFC-7539 conformity (notably, specifying
the new counter interface).

As a bonus, both the C program and the Cryptol program come equipped
with a test harness matching the RFC test vectors.

Tested with Clang 3.6.2 and SAW 0.2 on amd64/Linux.

Signed-off-by: Austin Seipp <aseipp@pobox.com>