Make divModNat a primitive
It would be nice to make addNat, subNat, mulNat and others primitives,
as well, but the source-level type checker doesn't know about them, so
it'll need to learn a little more before it'll work without
pattern-matching-based definitions for those functions. None of the
stuff in the prelude requires evaluation of divModNat for type checking.