Support verification of external LLVM functions
It is now possible to use `llvm_verify` on functions that are declared
but not defined as long as simulation is not necessary (i.e., as long as
the specification uses `llvm_no_simulate`). This is useful for
specifying the behavior of code not included in the system being
verified.