Add skeleton of `llvm_allocate` command
This will eventually specify that the method being verified (or
overridden) will, as part of its behavior, allocate a new chunk of
memory, of the type of the given location, and store the pointer to
the new memory in the given location. For now, however, it just
generates an error if you try to use it.