Add limited support for LLVM allocation specs
The `llvm_allocates` function specifies that the function allocates a
new value and stores a pointer to it in the given location. It is
particularly useful in unverified overrides for built-in constructor
functions.

It is more useful in overrides than in verification. In particular, it
does not currently check that such an allocation actually happened
during symbolic execution of the specified function. When used as an
override, however, it causes the allocation to happen.