Implement primitive `crucible_elem :: SetupValue -> Int -> SetupValue`.
This works basically like a getelementpointer operation in LLVM.
The first argument should be a SetupValue representing a pointer
to a struct or an array. The second argument is a field number
(for struct pointers) or an array index (for array pointers).
(All indexing is 0-based.) The resulting value is a pointer to
the appropriate element of the struct or array.