Don't use sbvSolveBasic on terms with ExtCns
When attempting to convert SAWCore Term values into concrete integers to
feed into Crucible, don't bother if the Term contains any ExtCns
subterms. Such terms tend to be functions of the symbolic input
variables, and therefore unlikely to evaluate to constants, and trying
to do so can take quite a lot of time. In some cases, the ExtCns
subterms might be simplified away, but this seems like a reasonable
compromise.