Make salsa-crucible.saw work completely. Fixes #189.
The idea is to avoid using crucible_points_to in the precondition section
if the function might be passed a pointer to uninitialized memory.
Pointer arguments that are used for output only should use just
crucible_alloc in the precondition section; crucible_points_to should
be used in the postcondition section only.