Fix #207 - terminate w/ non-zero when `crucible_llvm_verify` fails
Terminating the SAW binary seems a little harsh for the long run, but
for now there is no other way to programmatically determine the result
of a crucible verification short of parsing the stdout.  This change
also brings the behavior inline with `llvm_verify`.

I suggest we revert this patch once issue #208 is closed.