Pull in crucible-llvm pointer refactoring.  A few changes need to be made on the SAWScript side to account for the new bitvector representation. The main change is that, when translating `LLVMVal` values to SAWCore terms, the translation becomes type-directed, based on the Cryptol type. At bitvector types, the underlying bitvector value is injected into the LLVM pointer/bitvector type.
For now, there is no way to directly represent or refer to LLVM pointer values
in quoted Cryptol fragments; however, using abstract types in Cryptol, it
ought to be fairly straightforward to add the ability to refer to pointer
variables.  This would make it easier to specify values in SAWScript.