Make `checkMemTypeCompatibility` a bit more liberal about what it will accept.
It turns out that all the lower levels of the SAW/Crucible verification
stack will consider a value to be memory-layout compatible with a
single-field struct containing that value.  However, this needs to
be added as a special case to the procedure that check for memory
compatability of points-to and equality predicates in the SAWScript
front-end.