Switch llvm-pretty back to GaloisInc fork
Recommend swtiching GaloisInc/saw-script back to referencing its own fork of llvm-pretty. To do this, we'll need to merge the changes from elliottt.

This is desirable as it keeps all dependencies for SAW in one place. This makes it easier to manage the mirror of the Galois github that we have recently set up.