Fix build with GHC 8.4.1
Miscellaneous changes to get builds to work with GHC 8.4.1. Doing a full
build requires waiting for new Hackage versions of a couple of packages,
though. To make use of these changes, for now, you’ll have to clone the
latest git versions of a few packages and instruct the build to use
those versions instead of the Hackage ones.