Use Setup.hs instead of TH for commit version
In order to include the commit hash in the compiled executable, we used
to depend on some Template Haskell to run git at compile time. This code
required loading the static library for ABC into GHCi to run the
Template Haskell, which has always been brittle. Now, we generate the
module that contains the commit hash from a custom Setup.hs.