Handle recursive calls properly in scDefRewriteRules.
Currently this is implemented by calling rewriteSharedTerm
with a flipped version of the original rewrite rule. However
this would fail if the original rewrite rule had variables
on the left-hand-side that did not occur on the right. It
would be better to have a more flexible alternative to
scReduceRecursor, so that we could directly instantiate the
recursive calls the way we want.