More multiple class support
Moved the JVMContext to a new "JVMTranslation" structure in the
TopLevelRW context. Furthermore, the JVMContext is maintained
throughout class translations/extractions so that methods in all
classes can be found during simulation. Would be better to do this
a bit more lazily, but right now just trying to get it to work.