Add directory-locals so that intero-mode works correctly
All invocations of intero-mode in this directory use the target saw-script:lib
to load up the project. This prevents GHCi's linker from erroring out as
it (wrongly) attempts to load up abcBridge, which is not necessary for debugging
saw-script itself. This does mean that the saw-script top-level executable will
not have intero support, but that was true anyhow because it depends on
abcBridge.