Fix bug in p384_ec_mul_aux.
(This is a fix for #305, but now the ecdsa proof script fails.)

We need to mask the low 5 bits of the shift amount to match the
java behavior. From the JVM spec description of `ishl`:

"An int result is calculated by shifting value1 left by s bit positions,
where s is the value of the low 5 bits of value2. ... The shift distance
actually used is always in the range 0 to 31 inclusive, as if value2
were subjected to a bitwise logical AND with the mask value 0x1f."