llvm: don't initialize globals by default
Instead, add crucible_global_initializer.

The current behavior of globals is unsound (ref: #203). Instead,
we provide an option for the user to initialize globals as necessary.

It would be best to have more informative error messages when the
symbolic simulator accesses invalid memory because a global hasn't
been initialized.

I initially assumed that `SetupGlobalInitializer name` should have the
same type as `SetupGlobal name`. However, `SetupGlobal` adds a pointer
wrapper (as it should). Instead, we now directly translate the type
of the global's initializer from the LLVM AST.