Rework, again, how SAW handles multiple override specifications.
This time, the task is made simpler by first splitting the override
specification handling code into two pieces: one which performs structural
matching on the arguments and input memory state, and calculates
preconditions; and a second phase which updates the memory state,
computes return values and postconditions.

This way, we can directly compute the precondition predicates without
having to first be in the context of a symbolic branch.  Now we
build an if-then-else tree based on the precondition values, and
compute the postcondition state inside the body of the branch.
A final default case fails, which forces one of the provided
specifications to be selected.

The overall result is exactly what we want: the first override whose
preconditions are satisfied is selected, it's postconditions are assumed
and return values computed.  Crucible's path merging takes care of all
the bookeeping and no funny tricks are required.

Fixes issue #303.
Fixes issue #306.