Translate Cryptol record types to saw-core tuples.
NOTE: This means that the mapping of cryptol types to saw-core types
is no longer injective, so the types returned by `scCryptolType` and
`mkTypedTerm` may not match the original type of a term translated
from cryptol. If this is a problem, the use of `scCryptolType` and
`mkTypedTerm` should be avoided.