Use Data.Parameterized.Nonce instead of Control.Monad.Supply
Fixes #354, text from that issue:

monad-supply doesn't build with the latest GHC. The underlying issue is an
incomplete pattern match which fails with MonadFailDesugaring, and which will
take a bit of an API redesign to fix (so I'm guessing it won't happen too soon).
It also doesn't seem to be a widely used package (hackage reverse deps).