Update saw-script to use the latest macaw-symbolic
This is used for the x86 verification code.  There are a few API changes in
macaw symbolic that affected this code (largely renames), but a few functions
are now replaced with a cleaner API.

This change got caught up in some crucible-llvm changes (for the memory model)
that are also taken care of.