Implement uninterpreted functions.
Previously `freshConstant` was used to create a fresh variable for
each application of an uninterpreted function to a set of arguments;
the arguments themselves were discarded.

Now we use `freshTotalUninterpFn` to create a symbolic function that
is then applied to the list of arguments.

NOTE: This is still not correct, because we need to reuse the same
symbolic functions when we use the same uninterpreted function multiple
times at the same type.