Simplify ProofGoal representation; fix negation convention in backends.
As of this commit, we now treat every ProofGoal as having universally-
quantified variables. The `sat` command creates a universally-quantified
goal by logically negating the given predicate. As a result the solver
backends don't need to know whether the current proof goal came from a
`prove` or a `sat` command; they should behave the same in either case.

Fixes #322.