Prove all static variants instead of comparing
This avoids problems when various forms of metadata or duplicated static
globals make syntactic comparison of function bodies insufficient. It
may take slightly longer, but duplicated static functions tend to be
small.

This also changes the matching logic for static function names. Now two
names match if they are either equal or if one is a prefix of the other
and the remaining characters consist of a single `.` followed by any
number of digits.