First cut at model counting commands
The two new commands, `approxmc` and `sharpSAT`, use the command-line
programs of the same names to calculate the approximate and exact number
of solutions to a Term, respectively. The number of solutions is
relative to the CNF encoding, and is always larger than the number of
inputs to the SAWCore function itself.