Check bitvector types in `crucible_llvm_extract`
Because `crucible-llvm` merges integer and pointer types, it is not
sufficient to look at Crucible types to determine whether a CFG coming
from LLVM takes pointer arguments or returns a pointer. Instead, we have
to look at the original LLVM types. This commit does that, and produces
an error if `crucible_llvm_extract` is used with a function that has
poiinter types anywhere in its signature.

Fixes #521.