Speed up creation of uninterpreted functions.
Previously we accumulated a list of arguments to each uninterpreted
function, and then inferred their types by mapping `exprType` over
the argument list. It turns out that this can be somewhat expensive,
and when we ask for an uninterpreted function with a compound result
type, we create many functions with the same argument list and re-
compute the same argument types many times, multiplying the expense.

Now we compute the list of argument types in parallel with the arguments
themselves, which saves a great deal of time when uninterpreted functions
on very large types (e.g. [4096][8] -> [4096][8]) are used.