Use invariant for datatype `LLVMModule` to permit fast equality test.
The invariant is that the `LLVMModule` value should have been created
by the `loadLLVMModule` function, so that all of the components (the
file path, the parsed AST, and the translated Crucible CFG) all agree.

Also replace the lenses for the field components with ordinary
projection functions, as we should never want to modify any of the
record fields of an `LLVMModule`, only read them.