Add superclass instance computation.
We arrange it so every class dictionary contains it's superclass
dictionaries as a field, which can then be projected.  In `proveProp`,
whenever we get a class dictionary as an assumption, we saturate the
prop environment with all avaliable superclasses.  These are accessed
via a chain of field projections as necessary.