First semi-complete verification summarizer
This definitely needs to include a lot more information, but I think it
puts most of the necessary plumbing in place.

We should adapt the `SolverStats` tracking to include proofs done by
rewriting and `trivial`, as well. Currently, proofs based entirely on
rewriting are printed as axioms, which isn’t right.