Remove unused (and incorrect) case in function `baseCryptolType`.
This function is used to compute the cryptol type corresponding to
a crucible base type that may appear in a crucible CFG that is to be
extracted to a SAW function. Crucible array types are never actually
used in such cases, and if they were, they would not translate to
n-ary function types.