Add argo submodule and build (#853)
This commit moves the `saw-remote-api` package from the `argo` repository to the `saw-script` repository and adds it to the test suite. Some of the tests currently fail in ways due entirely to CI configuration, however, and are marked as `continue-on-error` to avoid holding up other work. We should fix those soon.