Add structured name support to SAWCore.  This involves a new `NameInfo` datatype that is attached to `ExtCns` (and thus `Constant`) nodes.
In addition, we implement a more organized naming environment, and a
notion of abolute names, which are intended to be a unique "name of
last resort".  Each name that doesn't derive from a SAWCore module is
required to have a unique URI.  Dynamically-created values incorporate
the unique VarIndex into their name, whereas concepts from, e.g.,
Cryptol modules have URIs that are derived from their fully-qualified
name.

In addition to the absolute name, imported names may have a collection
of aliases, which are the more typical way to refer to the name.
All the aliases for a name are put into a naming environment, which
map strings arising from user input, etc, into a set of names they
might refer to.  When the set contains more than one "true" name,
the identifier is ambiguous, and we can report all the names to which
that identifier might refer, as well as unambiguous alternate identifiers
for each name.