add a check to detect whether proofs are vacuous
If the user adds contradicting assumptions to their script, saw-script
will conclude that anything can be true.  This gives the user a false
sense of having proved their specification, when really, they have
proved nothing.

This checks whether the current context implies falsehood, in which case
a warning is displayed to the user.