Modify JVM class loader to avoid redundant map union operations.
The monoid append operation on JVMContext is rather inefficient,
because it involves unions on maps with keys that are expensive
to compare. In this case the append operation was completely
redundant, because the old context was being combined with a new
one that was the same as the old one extended with some new entries.