Extend `jvm_field_is` to support type-annotated field names.
JVM allows classes to have multiple fields with the same name
but different types. (Although it is unlikely to come across
such fields in programs compiled from Java, which disallows this.)

This change allows saw-script users to disambiguate this case
by adding a type annotation to a field name using type descriptor
notation. For example, "x:J" specifies a field named "x" of type
`long`. (This is the same field syntax used in the disassembly
output from `javap -c`.)