Make `jvm_field_is` and `jvm_elem_is` check type compatibility.
Fixes #904.

The check is performed at the spec-construction phase, as opposed
to during simulation; this means that type mismatches are caught
with either `crucible_jvm_verify` or `crucible_jvm_unsafe_assume_spec`.

The check uses the rather weak notion of register-compatibility as
the criterion. This means that any two reference types are considered
compatible, for example. It may be worth considering switching to
a stronger type compatibility relation at a later time.