Heapster bound names (#29)
* Translate bound names so as to avoid shadowing or using reserved keywords.

* Revert "bug fix: binders with the same string names were causing unintended shadowing"

This reverts commit 2e3c63d37919939db204c52108ce7546866f14e9.

The same bug was already fixed in PR #12.

* Fix `translatePi` so that it properly translates bound variables.

This patch fixes an oversight in the #12 PR. Previously it
incorrectly failed to use the name-collision-avoiding
`translateLocalIdent` function, and it similarly failed to
update the `localEnvironment` with the bound name.

* Fix incorrect local variable environment when translating lambdas.

This patch fixes another problem that was missed in PR #12.
Previously when translating the body of a lambda, the translator
would populate the local variable environment with the *old*
variable names instead of using the new ones. This was causing
incorrect output when a lambda variable had been renamed.

* Fix incorrect local variable environment when translating terms.

This patch fixes a few more instances of redundant and/or wrong
manipulations of the local variable environment when translating
terms. Most of these changes were adapted from the wip-heapster
branch (revision 2e3c63d3).