Refine `JVMPointsTo` datatype to store `AllocIndex` of left-hand side.
Previously it represented the left-hand side with an arbitrary
`SetupValue`; however, the only valid `SetupValue`s for this purpose
are `SetupVar` applied to an `AllocIndex`. This change lets us
simplify code for the simulation phase, in exchange for doing more
error checking during the setup phase (which is also a good thing).

The early checking eliminates a couple of the known false positives
from #938.