Modify `testResolved` to detect partially-overlapping points-tos.
Previously `testResolved` would detect a later points-to that pointed
entirely inside the region of an earlier points-to, but would not
detect an overlap if the points-tos were done in the opposite order.

This change makes the points-to comparison symmetric, so overlapping
points-to declarations are detected no matter what order they appear
in a spec.

This fixes one of the remaining known false positives from #938.