Add strictness to fix a space leak in crucible_llvm_unsafe_assume_spec.
Due to laziness, some thunks inside the MethodSpec datatype had been
holding references to the `LLVMCrucibleContext`, whose `ccLLVMGlobals`
field contains the crucible-llvm MemImpl data structure, which in turn
contains copies of all const globals. A separate copy of this data
structure was retained for every override value.

Profiling showed that unevaluated thunks resided in the `allocSpecAlign`
field of `LLVMAllocSpec`, and also in the `csVarTypeNames` field of
`StateSpec`, so these areas have been made more strict.

This patch is somewhat related to #850, although this just fixes the space
leak, and doesn't remove any of the unnecessary or duplicated computation.