Adjust jvm_verify type-based method disambiguation syntax.
The syntax for appending a method type descriptor to a method name
to disambiguate between similarly-named methods in `jmv_verify`
is modified to require a colon between the method name and the
descriptor. For example, "foo:(J)V" instead of "foo(J)V".

This makes the descriptor syntax consistent with that used to
disambiguate field names for `jvm_field_is`. It also makes the
syntax consistent with that shown by the `javap` disassembler.