Make executable finding more robust in CI
When changing the version of a package, builds for the old version are
cached so there will likely be two versions of all the executables. We
just want one of them, and the newest one. This seems to do the trick
for now, but I'm eagerly awaiting the time when Cabal will do this for
us.