Switch from `GlobalDef` to `Constant` for definitions in .sawcore files.
This implements most of #43.

The saw-core parser/typechecker now uses `scConstant` to represent
defined constants. The `GlobalDef` constructor is still used to
represent primitives and axioms declared in .sawcore files.

The change is done in a backward-compatible way: The SharedContext
now contains an environment mapping from `Ident`s to `Term`s, so
`scGlobalDef` can be used in the same way as before (although the
checking for validity of names is more strict now).

The recognizer `asGlobalDef` is also made backward compatible: It
now recognizes either `GlobalDef` constants or `Constant` terms
that have a name made from an `Ident`.

Finally, the simulator has been updated to accommodate backends that
implement "primitives" that are actually constants with overridden
definitions: In the uninterpreted-function handler, the prims map is
also consulted to see if the `Constant` under consideration was
a converted `GlobalDef` that has another implementation.