Update examples/java/branchsat.saw to use new jvm_verify commands.
This example is supposed to demonstrate the use of path sat checking.
With the old java_verify command, the original example required
path sat checking enabled even if the input value is concrete, while
jvm_verify does not, so the example is no longer a good demonstration
of the need for path sat checking. However, a fully symbolic input
with a symbolic precondition does not yet terminate even with path
sat checking enabled. We'll leave a "FIXME" comment reminding us to
update the example when we enhance jvm_verify to handle that case.