CI quickfix: Pin libBF==0.5.1 in cabal.GHC-8.6.5.config
For some reason, `libBF`'s version was not being pinned in
`cabal.GHC-8.6.5-config`, causing CI to pick a build plan with `libBF-0.6`.
This leads to the build error observed in
https://github.com/GaloisInc/cryptol/pull/1055#issuecomment-771649363.

A quick fix for the issue to make sure that `libBF-0.5.1` is being picked
in the 8.6.5 build plan, just like it is for other versions of GHC that CI
tests. To this end, I manually inserted `any.libBF ==0.5.1` into
`cabal.GHC-8.6.5-config`.