Bump cryptol to include support for libBF>=0.6
libBF is not included in most of cryptol's freeze files, causing builds
of cryptol at the prior commit to start failing once libBF 0.6 was
released. Cryptol is built from the submodule by nightly and for bindist
packaging. Bumping the submodule to 6e3c0816 includes support for libBF
0.6 resolving these build failures (for now).

The freeze file updates also remove the need to use allow-newer for the
build to work.