Eschew -Weverything in favor of -Wall
`-Weverything` enables absolutely every single warning that GHC has in its
arsenal, which is usually too much. As discussed in #1083, this replaces
`-Weverything` in favor of `-Wall`, which enables a more reasonable subset of
warnings. This means that we don't have to explicitly disable as many warnings
(which would otherwise be enabled by `-Weverything`) as before.

Fixes #1083.