Introduce llvm_alias, make llvm_struct a synonym for llvm_alias
This introduces the `llvm_alias` function and makes `llvm_struct` a synonym for
`llvm_alias`. For now, `llvm_struct` continues to be available without a
deprecation notice. In the future, we may want to deprecate `llvm_struct` in
favor of `llvm_alias`, as having two separate functions named `llvm_struct` and
`llvm_struct_type` could invite confusion. See the discussion in #1085.