Continue making the SAWScript proof abstractions stronger.
The `Theorem` and `ProofState` data types are now abstract,
and changes in proof state are mostly mediated by the
new `Tactic` concept.  Theorems now contain `Evidence`
which explains the reasoning for why we accept a particular
proposition as true. Proof states now explicitly construct
evidence as part of tactic execution, and evidence is checked
against the proposition to be proved when finishing a proof.

Forms of evidence we accept are:

1) Explicit proof terms, according to types-as-propositions
2) Black-box propositions either directly admitted,
   random checked via quickcheck or proved by solvers
3) Rewriting, unfolding and conversion steps on goals
4) Sequent calculus reasoning rules (theorem application, cut,
   introduction of local hypotheses, etc).

This isolates the core soundness concepts to the "SAWScript.Proof"
module, where the key invariants can be maintained relatively
easily.