Remove unused code from `evalGlobal` in `Simulator.hs`.
Since #118, we no longer need to rely on `TypedDef`s from the
module map to evaluate global definitions, because they are
now represented as `Constant` terms that include the body of the
definition directly in the term.

This change reduces the startup cost for running a simulator
backend, because it reduces the size of the `globals` map that
is constructed for each call to `evalGlobal`.