Reimplement `SimulatorConfig` and `evalGlobal` to avoid rebuilding maps.
Previously `evalGlobal` would build a new `HashMap` containing all
primitives and all data constructors. Now the `SimulatorConfig` is
updated such that rebuilding such a map is not necessary; instead
the provided table of primitives is used directly by the evaluator.
This significantly reduces the startup cost of calling
`evalSharedTerm` in various backends.