Compare pointers that are not valid LLVM pointers.
Comparing the offsets as unsigned bitvectors is not sound, because of
overflow (e.g. `base - 1` is less than `base`, but -1 is not less than 0
when compared as unsigned). It is safe to allow a small negative offset,
because each pointer base is mapped to an address that is not in the
first page, which is never mapped on X86_64 Linux.