Rework the `runProofScript` method.
It now accepts a `ProofGoal` and computes a `ProofResult`, which
is either a completed proof, a counterexample to a subgoal of the
proof, or an incomplete proof. All the primitives accepting
`ProofScript` arguments are now work independently of what the
proof script action computes.  In addition, verification methods
are more careful to check that all proof obligations are discharged,
whereas previously they were not.